refactor: simpler models layout - custom navigation drawer - custom toolbar - extracted into subcomponents - tweaks to pages - adjusted global styles
