+++ src/components/navigation/index.ts
@@ -1,0 +1,7 @@
+import NavGroup from './nav-group';
+import NavLink from './nav-link';
+
+export * from './nav-group';
+export * from './nav-link';
+
+export { NavGroup, NavLink };
