+++ src/pages/_app.tsx
@@ -9,11 +9,6 @@ import store from '../store';
 import '../styles/globals.css';
 
 function App({ Component, pageProps }: AppProps): ReactElement {
-  useEffect(function removeServerSideCSS() {
-    const jssStyles = document.querySelector('#jss-server-side');
-    if (jssStyles) jssStyles.parentElement?.removeChild(jssStyles);
-  }, []);
-
   return (
     <Provider store={store}>
       <ThemeProvider theme={theme}>
