+++ src/styles/globals.css
@@ -7,3 +7,12 @@ a {
   color: inherit;
   text-decoration: none;
 }
+
+html, body {
+  height: 100%;
+}
+
+#__next {
+  width: 100%;
+  height: 100%;
+}
