+++ src/styles/theme.ts
@@ -1,5 +1,8 @@
 import { createMuiTheme } from '@material-ui/core';
 
+const defaultTheme = createMuiTheme();
+
 export const theme = createMuiTheme({
-  spacing: (factor: number) => `${0.25 * factor}rem`,
+  spacing: (value: number) => defaultTheme.typography.pxToRem(value / 0.25),
+  // spacing: (value: number) => `${0.25 * value}rem`,
 });
